+++ packages/website-eslint/src/mock/util.js
@@ -4,4 +4,4 @@ util.inspect = function (value) {
   return value;
 };
 
-export default util;
+module.exports = util;