+++ packages/website/src/components/linter/config.ts
@@ -1,6 +1,9 @@
-import type { ParseSettings } from '@typescript-eslint/typescript-estree/use-at-your-own-risk/parseSettings';
+import type { ParseSettings } from '@typescript-eslint/typescript-estree/use-at-your-own-risk';
+import type { TSESLint } from '@typescript-eslint/utils';
 
-export const parseSettings: ParseSettings = {
+export const PARSER_NAME = '@typescript-eslint/parser';
+
+export const defaultParseSettings: ParseSettings = {
   allowInvalidAST: false,
   code: '',
   codeFullText: '',
@@ -13,7 +16,7 @@ export const parseSettings: ParseSettings = {
   EXPERIMENTAL_useSourceOfProjectReferenceRedirect: false,
   extraFileExtensions: [],
   filePath: '',
-  jsx: false,
+  jsx: true,
   loc: true,
   log: console.log,
   preserveNodeMaps: true,
@@ -25,4 +28,18 @@ export const parseSettings: ParseSettings = {
   tokens: [],
   tsconfigMatchCache: new Map(),
   tsconfigRootDir: '/',
+};
+
+export const defaultEslintConfig: TSESLint.Linter.Config = {
+  parser: PARSER_NAME,
+  parserOptions: {
+    ecmaFeatures: {
+      jsx: true,
+      globalReturn: false,
+    },
+    ecmaVersion: 'latest',
+    project: ['./tsconfig.json'],
+    sourceType: 'module',
+  },
+  rules: {},
 };